+++ R/ExposureProportions.R
@@ -4,17 +4,38 @@
 #'
 #' @param cancer.type A character string.
 #'
-#' @param all.sigs An optional matrix of known signatures, with column names being signatures ids.
+#' @param all.sigs An optional matrix of known signatures, 
+#'    with column names being signatures ids. Only used to drop 
+#'    signatures not present in \code{all.sigs}.
 #'
-#' @param drop.sigs.no.info If TRUE, drop any not present in the column names of \code{all.sigs}.
+#' @param drop.sigs.no.info If TRUE, drop signatures any not present in 
+#'    the column names of \code{all.sigs}.
 #'    There are some signatures that do not have SBS192 versions, including SBS29.
+#'    
+#' @param must.include A character vector of signature IDs that
+#'    must be included, even if they have not previously been
+#'    observed in that cancer type. The associated proportion is
+#'    specified by \code{new.sig.prop}.
+#'    
+#' @param must.include.prop The value used for the expected proportion of
+#'    signatures in \code{must.include} but not previously observed
+#'    in the given \code{cancer.type}.
 #'
 #' @export
 #'
-#' @return A numerical vector of the proportion of tumors of type \code{cancer.type} with each signature
-#'   for those signatures observed in \code{cancer.type}.The names are the signature ids.
+#' @return A numerical vector of the proportion of 
+#'   tumors of type \code{cancer.type} with each signature
+#'   for those signatures observed in \code{cancer.type}. 
+#'   The names are the signature ids.
 #'
-ExposureProportions <- function(mutation.type, cancer.type, all.sigs = NULL, drop.sigs.no.info = TRUE) {
+ExposureProportions <- function(
+  mutation.type, 
+  cancer.type, 
+  all.sigs          = NULL, 
+  drop.sigs.no.info = TRUE,
+  must.include      = character(),
+  must.include.prop  = 0.1
+  ) {
 
   mtype <- ifelse(mutation.type == "SBS192", "SBS96", mutation.type)
 
@@ -27,6 +48,13 @@ ExposureProportions <- function(mutation.type, cancer.type, all.sigs = NULL, dro
   sigs.prop <- unlist(sigs.prop[ , 2])
   names(sigs.prop) <- sig.names
 
+  other.prop.sig.names <- setdiff(must.include, sig.names)
+  sig.names <- c(sig.names, other.prop.sig.names)
+  other.props <-
+    rep.int(must.include.prop, times = length(other.prop.sig.names))
+  names(other.props) <- other.prop.sig.names
+  sigs.prop <- c(sigs.prop, other.props)
+  
   if (!is.null(all.sigs)) {
     sigs.no.info <- setdiff(sig.names, colnames(all.sigs))
 
@@ -36,8 +64,10 @@ ExposureProportions <- function(mutation.type, cancer.type, all.sigs = NULL, dro
       sigs.prop <- sigs.prop[-(which(names(sigs.prop) == zz))]
     }
 
-    if (!drop.sigs.no.info) stop("There were signaures with no proporiton information")
+    if (!drop.sigs.no.info) 
+      stop("There were signaures with no proporiton information")
   }
+  
   return(sigs.prop)
 }